Added some JS & CSS Boilerplate. I think it starts to make sense to move these to external locations and repositories.
The problem is that they're all small components that would on their own
make up for a gigantic amount of extra management work and some of them
are really tightly woven into the system itself.

Probably the best course of action is to maintain the files in a regular
fashion at first (with file managers) and start isolating the 
repositories once I have a clearer concept on their workings.